project/apriltag/from zhaoxin/table_guidance_project_v3/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "table-guidance"
version = "0.2.1"
description = "AprilTag + FastSAM segmentation -> planar coordinate guidance on a table."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.23",
  "opencv-python>=4.7",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
apriltag = [
  "pupil-apriltags>=1.0.4",
]
fastsam = [
  "ultralytics>=8.0.0",
  # torch is required by ultralytics; install the correct build for your platform
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
